feat : BlackPutRule 구현
