+++ src/main/kotlin/omok/model/PutRule.kt
@@ -1,0 +1,30 @@
+package omok.model
+
+import lib.ark.ark.ArkFourFourRule
+import lib.ark.ark.ArkOverLineRule
+import lib.ark.ark.ArkThreeThreeRule
+import omok.mapper.toArkOmokBoard
+import omok.mapper.toArkOmokPoint
+
+interface PutRule {
+    fun canPut(stone: OmokStone, board: Board): Boolean
+}
+
+object BlackPutRule : PutRule {
+    override fun canPut(stone: OmokStone, board: Board): Boolean {
+        val arkBoard = board.toArkOmokBoard()
+        val arkPoint = stone.position.toArkOmokPoint()
+        val isNotFourFour = ArkFourFourRule.validate(arkBoard, arkPoint).not()
+        val isNotThreeThree = ArkThreeThreeRule.validate(arkBoard, arkPoint).not()
+        val isNotJangMok = ArkOverLineRule.validate(arkBoard, arkPoint).not()
+        return isNotFourFour && isNotThreeThree && isNotJangMok && WhiteCanPutRule.canPut(stone, board).not()
+    }
+}
+
+object WhiteCanPutRule : PutRule {
+    override fun canPut(stone: OmokStone, board: Board): Boolean {
+        val isEmptyPosition = board.isEmptyPosition(stone)
+        val isInRange = board.isInRange(stone)
+        return isEmptyPosition && isInRange
+    }
+}
