+++ src/test/kotlin/omok/BoardTest.kt
@@ -24,7 +24,7 @@ class BoardTest {
             )
         val expected = createOmokStone(x = 2, y = 2, color = StoneColor.WHITE)
         // when
-        val actual = board.last()
+        val actual = board.lastOrNull()
         // then
         actual shouldBe expected
     }
